backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.3
pydantic==2.9.0
python-dotenv==1.0.1
cassandra-driver==3.29.2
requests==2.32.3
email-validator==2.1.1
regex==2024.7.24
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
textblob==0.18.0
pyspellchecker==0.8.1
google-api-python-client==2.142.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
psycopg[binary]==3.2.12
joblib==1.4.2
pytest==8.3.2
httpx==0.27.0
black==24.8.0
isort==5.13.2

